LayoutTests/fast/css/webkit-column-span-calculated-value.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testDiv"></div>
<script>
description("Tests assigning a calculated value to '-webkit-column-span' CSS property.");

var testDiv = document.getElementById("testDiv");

shouldBeEmptyString("testDiv.style['-webkit-column-span']");
evalAndLog("testDiv.style['-webkit-column-span'] = 'calc(3 - 2)'");
shouldBeEqualToString("testDiv.style['-webkit-column-span']", "1");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('-webkit-column-span')", "none");
testDiv.style['-webkit-column-span'] = '';

shouldBeEmptyString("testDiv.style['-webkit-column-span']");
evalAndLog("testDiv.style['-webkit-column-span'] = '1'");
shouldBeEqualToString("testDiv.style['-webkit-column-span']", "1");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('-webkit-column-span')", "none");
testDiv.style['-webkit-column-span'] = '';

shouldBeEmptyString("testDiv.style['column-span']");
evalAndLog("testDiv.style['column-span'] = 'calc(3 - 2)'");
shouldBeEqualToString("testDiv.style['column-span']", "1");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('column-span')", "none");
testDiv.style['column-span'] = '';

shouldBeEmptyString("testDiv.style['-webkit-column-span']");
evalAndLog("testDiv.style['-webkit-column-span'] = 'calc(3 * 2)'");
shouldBeEmptyString("testDiv.style['-webkit-column-span']")

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
