LayoutTests/fast/css/webkit-column-span-calculated-value-expected.txt

Tests assigning a calculated value to '-webkit-column-span' CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['-webkit-column-span'] is ""
testDiv.style['-webkit-column-span'] = 'calc(3 - 2)'
PASS testDiv.style['-webkit-column-span'] is "1"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-column-span') is "none"
PASS testDiv.style['-webkit-column-span'] is ""
testDiv.style['-webkit-column-span'] = '1'
PASS testDiv.style['-webkit-column-span'] is "1"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-column-span') is "none"
PASS testDiv.style['column-span'] is ""
testDiv.style['column-span'] = 'calc(3 - 2)'
PASS testDiv.style['column-span'] is "1"
PASS window.getComputedStyle(testDiv).getPropertyValue('column-span') is "none"
PASS testDiv.style['-webkit-column-span'] is ""
testDiv.style['-webkit-column-span'] = 'calc(3 * 2)'
PASS testDiv.style['-webkit-column-span'] is ""
PASS successfullyParsed is true

TEST COMPLETE
